requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
asyncpg==0.29.0
ccxt==4.4.6
pandas==2.2.3
pandas-ta==0.3.14b0
numpy==1.26.4
aiohttp==3.10.5
openai==1.51.0
python-telegram-bot==21.6
python-dotenv==1.0.1
slowapi==0.1.9
